Handshake client now using verifier
