+++ src/SlackConnector/Connections/ConnectionFactory.cs
@@ -1,5 +1,4 @@
-﻿using SlackConnector.Connections.Handshaking;
-using SlackConnector.Connections.Messaging;
+﻿using SlackConnector.Connections.Messaging;
 using SlackConnector.Connections.Sockets;
 using SlackConnector.Connections.Sockets.Messages.Inbound;
