+++ src/SlackConnector/SlackConnector.csproj
@@ -58,8 +58,8 @@
     <Compile Include="BotHelpers\IMentionDetector.cs" />
     <Compile Include="BotHelpers\IChatHubInterpreter.cs" />
     <Compile Include="Connections\ConnectionFactory.cs" />
-    <Compile Include="Connections\Handshaking\HandshakeClient.cs" />
-    <Compile Include="Connections\Handshaking\IHandshakeClient.cs" />
+    <Compile Include="Connections\Messaging\HandshakeClient.cs" />
+    <Compile Include="Connections\Messaging\IHandshakeClient.cs" />
     <Compile Include="Connections\Messaging\ChannelMessenger.cs" />
     <Compile Include="Connections\Messaging\IChannelMessenger.cs" />
     <Compile Include="Connections\Messaging\IChatMessenger.cs" />
